Migrated more mehfil related functonality to its own views, forms, and template folder. Also patched an exploit in signup funnel (unauth_forms)
